add screenOn window flag when handle onNewIntent callback
Today we only start screenOn Timer when alert activity was created
first time. However, if a new alert is re-launched,
instead of a new instance of the alert activity being started,
onNewIntent callback will be triggered instead. we should turn screenOn
timer as well in order to display the second alert when screen is off.
This will fix the issue that users did not ack the alert and receive
a new alert later, the second alerts will not display when screen is off.

Bug: 162707010
Test: Manual test on two back to back alert when screen is off.
Change-Id: Iba20d2aff1b53e4071eb03c041e434f54384dcf9
(cherry picked from commit 01baad24901b9840fbb91d3b8346ea80cdd4c440)