Fix ANR by mocking activitymanager during unittest
In case of manifest broadcastreceiver, mocking the ActivityManager doesn't prevent the real ActivityManager from delivering the broadcast.
As a result, the real ActivityManager doesn't receive the completion signal and incorrectly assumes the receiver is still processing the broadcast, leading to an ANR.
So we need to Mock the behavior of IActivityManager.finishReceiver() to forward any calls to the actual ActivityManager service.

Bug: 350372237
Test: atest GoogleCellBroadcastReceiverUnitTests
Flag: TEST_ONLY
Change-Id: I1f250e324bb7f4666e030ef1a106ea4dc32a067c